pyproject.toml
[project]
name = "crew4ai-config"
version = "0.1.0"
description = "Pre-configured CrewAI crews for common software development and system administration tasks"
authors = [
    {name = "Dotfiles Team"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10,<3.11"
dependencies = [
    "crewai>=0.177.0",
    "crewai-tools>=0.69.0",
    "pyyaml>=6.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=4.0",
]

[tool.black]
line-length = 88
target-version = ['py310']